requirements.txt
# requirements.txt
pymed
haystack-ai
transformers
python-dotenv
streamlit